+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This change
 ## [2-beta29] - unreleased
 - updated deps
 - prevented unnecessary `def`s in generated test namespaces
+- changed the loading animation to an opacity change, to avoid the unnecessary page refresh
 
 ## [2-beta28] - 2025-01-19
 - added a `snippets` API namespace for easier editor integration - by @timothyprately - experimental